refactor: restructure CLI commands into separate files for improved organization and maintainability; add Go code quality guidelines to documentation
